Mon 16454 add alma 8 to centreon collect nightly tests (#506) (#519)
* enh(ci): Add alma8 to centreon-collect nightly

* Add job to robot-nightly workflow

* enable ci test branch

* Update robot nightly workflow

* add debug to get-packages

* Add release to github_env

* add delivery test

* Get pkg dir version name

* Fix url

* fix url

* make curl silent

* fix curl

* add mariadb client

* update get-packages

* disable nightly on alma 8

* update alma8 image

* Add MariaDB packages to alma8 dockerfile

* remove extra sudo

* update collect-test-robot.sh

* Remove duplicate Robot test job

* fix(ci/tests): attempt to upgrade grpc python library

* Install proper grpc version based on OS version id

* help cmake to find protoc

* disable delivery tests

* remove unused lines from actions

* Remove more unused lines

+++ ci/scripts/collect-test-robot.sh
@@ -32,9 +32,26 @@ echo "########################### Install Robot Framework ######################
 cd /src/tests/
 pip3 install -U robotframework robotframework-databaselibrary pymysql python-dateutil
 
-yum install "Development Tools" python3-devel -y
+yum groupinstall "Development Tools" -y
+yum install python3-devel -y
 
-pip3 install grpcio==1.33.2 grpcio_tools==1.33.2
+# Get OS version id
+VERSION_ID=$(grep '^VERSION_ID' /etc/os-release | sed -En 's/^VERSION_ID="([[:digit:]])\.[[:digit:]]"/\1/p')
+
+# Force version for el7 only
+if [ -f /etc/os-release ]; then
+    case "$VERSION_ID" in
+        7)
+            pip3 install grpcio==1.33.2 grpcio_tools==1.33.2
+            ;;
+        8)
+            pip3 install grpcio grpcio_tools
+            ;;
+        *)
+            echo "OS Version is neither 7 or 8"
+            ;;
+    esac
+fi
 
 ./init-proto.sh
